Removed props from the floating cart button and renamed correctly
